Standardize copyright includes, close layout divs, fix SQL error and deprecation notice in dashboard_shipping, fix payroll runs layout

+++ dashboard/listing_shipping_customers.php
@@ -4,7 +4,7 @@ use App\Core\DB;
 include('admin_elements/admin_header.php');
 $module = 'shipping_customers';
 $module_caption = 'Shipping Customer';
-$tbl_name = DB::SHIPPING_CUSTOMERS;
+$tbl_name = DB::CUSTOMERS;
 $error_message = '';
 $success_message = '';
 
@@ -37,9 +37,9 @@ $activeOrganizationId = dashboardRequireActiveOrganization();
 if (($action == "delete_$module" && !empty($id)) && granted('delete', $module_id)) {
 
     if (is_SuperAdmin()) {
-        $mysqli->query("DELETE FROM `$tbl_name` WHERE id=$id");
+        $mysqli->query("DELETE FROM `$tbl_name` WHERE id=$id AND entity_type='shipping'");
     } else {
-        $mysqli->query("DELETE FROM `$tbl_name` WHERE id=$id AND created_by ='". $session_user_id ."'");
+        $mysqli->query("DELETE FROM `$tbl_name` WHERE id=$id AND entity_type='shipping' AND created_by ='" . $session_user_id . "'");
     }
 
 
@@ -61,26 +61,26 @@ if (($action == "delete_$module" && !empty($id)) && granted('delete', $module_id
 <div class="content-wrapper">
 
     <!-- Page header -->
-    <div class="page-header page-header-light shadow">
-        <div class="page-header-content d-lg-flex border-top">
-            <div class="row mt-2">
-                <div class="col-lg-12">
-                    <h5 class="ms-2 mb-0"> <a href="listing_<?php echo $module; ?>.php" class="text-dark">All <?php echo ucwords(str_ireplace('_', " ", $module)); ?></a></h5>
-                </div>
-
-                <a href="#breadcrumb_elements" class="btn btn-light align-self-center collapsed d-lg-none border-transparent rounded-pill p-0 ms-auto" data-bs-toggle="collapse">
-                    <i class="ph-caret-down collapsible-indicator ph-sm m-1"></i>
-                </a>
+    <div class="page-header page-header-light shadow carriers-page-header">
+        <div class="page-header-content border-top py-2 px-3 carriers-page-header-content">
+            <div class="my-1">
+                <h1 class="h5 mb-0 d-inline-flex align-items-center gap-2">
+                    <a href="listing_<?php echo $module; ?>.php" class="text-dark">All <?php echo ucwords(str_ireplace('_', " ", $module)); ?></a>
+                    <?php if (!empty($pageHelpData)): ?>
+                        <button type="button" class="page-help-trigger-btn" data-bs-toggle="offcanvas" data-bs-target="#pageHelpPanel" title="How to use this page" aria-label="Page help">
+                            <i class="ph-question"></i>
+                        </button>
+                    <?php endif; ?>
+                </h1>
             </div>
 
-            <?php if (isset($module_id) && granted('create', $module_id)) { ?>
-                <!-- <div class="collapse d-lg-block ms-lg-auto" id="breadcrumb_elements">
-                    <div class="d-lg-flex mb-2 mb-lg-0">
-                        <button type="button" class="btn btn-primary btn-sm mt-1 mb-1" onclick="window.location.href='<?php echo $module; ?>.php?action=add_<?php echo $module; ?>';"><i class="ph-plus ph-sm me-2 opacity-75"></i>Add New</button>
-                    </div>
-                </div> -->
-            <?php } ?>
-
+            <div class="my-1">
+                <?php if (empty($hide_add_button) && isset($module_id) && isset($module) && granted('create', $module_id)) { ?>
+                    <!-- <a href="<?php echo $module; ?>.php" class="btn btn-primary btn-sm d-inline-flex align-items-center">
+                        <i class="ph-plus ph-sm me-2 opacity-75"></i>New
+                    </a> -->
+                <?php } ?>
+            </div>
         </div>
     </div>
     <!-- /page header -->
@@ -109,8 +109,8 @@ if (($action == "delete_$module" && !empty($id)) && granted('delete', $module_id
                     </div>
                 </div>
 
-<?php include('admin_elements/copyright.php'); ?>
     </div>
+    <?php include('admin_elements/copyright.php'); ?>
 </div>
 
 <script>
@@ -133,7 +133,5 @@ $(document).ready(function() {
     });
 });
 </script>
-
-<?php include('admin_elements/admin_footer.php'); ?>
 
 <?php include('admin_elements/admin_footer.php'); ?>
